Move primary interfaces out of the folder
